+++ internal/domain/ports/order_item_repository.go
@@ -1,3 +1,11 @@
 package ports
 
-type OrderItemRepository interface{}
+import (
+	"context"
+
+	"github.com/d-yanez/dys-sub-fsc-orders/internal/domain/entities"
+)
+
+type OrderItemRepository interface {
+	UpsertMany(ctx context.Context, items []entities.OrderItem) error
+}