+++ internal/domain/ports/order_repository.go
@@ -1,3 +1,11 @@
 package ports
 
-type OrderRepository interface{}
+import (
+	"context"
+
+	"github.com/d-yanez/dys-sub-fsc-orders/internal/domain/entities"
+)
+
+type OrderRepository interface {
+	Upsert(ctx context.Context, order entities.Order) error
+}